feat: improve setup
