+++ resources/my-header.mjs
@@ -1,0 +1,4 @@
+function MyHeader({ html }) {
+  return html`<style>h1 { color: red; }</style><h1><slot></slot></h1>`
+}
+
